+++ CHANGES.rst
@@ -10,6 +10,7 @@ Cubeviz
 
 - Added functionality to Collapse and Spectral Extraction plugins to save results to FITS file. [#2586]
 
+- Moment map plugin now supports linear per-spaxel continuum subtraction. [#2587]
 
 Imviz
 ^^^^^
@@ -25,6 +26,9 @@ Specviz2d
 
 API Changes
 -----------
+
+- ``width`` argument in Line Analysis plugin is renamed to ``continuum_width`` and ``width``
+  will be removed in a future release. [#2587]
 
 Cubeviz
 ^^^^^^^